use non-relative component imports and rename wc to wallet-connect
