+++ apps/web/src/components/wallet-transfers/success-stage.tsx
@@ -1,10 +1,10 @@
 import { CheckCircleIcon } from '@heroicons/react/outline'
 import useTranslation from 'next-translate/useTranslation'
 
-import Heading from '../heading'
-import LinkButton from '../link-button'
-
 import css from './wallet-transfers.module.css'
+
+import Heading from '@/components/heading'
+import LinkButton from '@/components/link-button'
 
 export interface SuccessStageProps {
   linkUrl: string
